Pull Brain management out of Pinky into BrainManager.
